hice algunos cambios

+++ src/main/java/com/uade/grupo4/backend_ecommerce/repository/entity/Profile.java
@@ -36,11 +36,11 @@ public class Profile {
         this.email = email;
     }
 
-    public long getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 }
